feat: tratamento de exceção

+++ src/main/java/io/github/dev_emanuelpereira/libraryapi/service/AutorService.java
@@ -1,8 +1,13 @@
 package io.github.dev_emanuelpereira.libraryapi.service;
 
+import io.github.dev_emanuelpereira.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.dev_emanuelpereira.libraryapi.model.Autor;
 import io.github.dev_emanuelpereira.libraryapi.repository.AutorRepository;
+import io.github.dev_emanuelpereira.libraryapi.repository.LivroRepository;
+import io.github.dev_emanuelpereira.libraryapi.validator.AutorValidator;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,8 +18,11 @@ import java.util.Optional;
 public class AutorService {
 
     private final AutorRepository autorRepository;
+    private final AutorValidator autorValidator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor) {
+        autorValidator.validarAutor(autor);
         return autorRepository.save(autor);
     }
 
@@ -23,10 +31,13 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir autores que possuem livros. Exclua os livros primeiro.");
+        }
         autorRepository.delete(autor);
     }
 
-    public List<Autor> pesquisa (String nome, String nacionalidade) {
+    /**public List<Autor> pesquisa (String nome, String nacionalidade) {
         if (nome != null && nacionalidade != null){
             return autorRepository.findByNomeAndNacionalidade(nome, nacionalidade);
         }
@@ -37,13 +48,34 @@ public class AutorService {
             return autorRepository.findByNacionalidade(nacionalidade);
         }
         return autorRepository.findAll();
+    }**/
+
+    public List<Autor> pesquisa(String nome, String nacionalidade) {
+
+        Autor autor = new Autor();
+        autor.setNome(nome);
+        autor.setNacionalidade(nacionalidade);
+
+        ExampleMatcher matcher = ExampleMatcher
+                .matching()
+                .withIgnoreNullValues()
+                .withIgnoreCase()
+                .withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING);
+        Example<Autor> autorExample = Example.of(autor, matcher);
+
+        return autorRepository.findAll(autorExample);
     }
 
     public void atualizar(Autor autor) {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar é necessario que autor exista");
         }
+    autorValidator.validarAutor(autor);
     autorRepository.save(autor);
+    }
+
+    public boolean possuiLivro (Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 
 }
